+++ mbo/container/internal/limited_ordered.h
@@ -178,7 +178,6 @@ class [[nodiscard]] LimitedOrdered {
     using value_type = LimitedOrdered::value_type;
     using pointer = LimitedOrdered::const_pointer;
     using reference = LimitedOrdered::const_reference;
-    using element_type = LimitedOrdered::value_type;
 
     constexpr const_iterator() noexcept : pos_(nullptr) {}  // Needed for STL
 
@@ -273,7 +272,6 @@ class [[nodiscard]] LimitedOrdered {
     using value_type = LimitedOrdered::value_type;
     using pointer = LimitedOrdered::pointer;
     using reference = LimitedOrdered::reference;
-    using element_type = LimitedOrdered::value_type;
 
     constexpr iterator() noexcept : pos_(nullptr) {}  // Needed for STL